--- Depth/Projectiles/DepthGlobalProjectile.cs
@@ -1,20 +1,0 @@
-using Microsoft.Xna.Framework;
-using System;
-using Terraria;
-using Terraria.ID;
-using Terraria.Localization;
-using Terraria.ModLoader;
-
-namespace Depth.Projectiles
-{
-    public class DepthGlobalProjectile : GlobalProjectile
-    {
-        public override bool InstancePerEntity
-        {
-            get { return true; }
-        }
-
-        public Item shotFrom = null;
-    }
-
-}
