--- Depth/Properties/launchSettings.json
@@ -1,14 +1,0 @@
-{
-  "profiles": {
-    "Terraria": {
-      "commandName": "Executable",
-      "executablePath": "$(tMLPath)",
-      "workingDirectory": "$(TerrariaSteamPath)"
-    },
-    "TerrariaServer": {
-      "commandName": "Executable",
-      "executablePath": "$(tMLServerPath)",
-      "workingDirectory": "$(TerrariaSteamPath)"
-    }
-  }
-}
