--- Depth/build.txt
@@ -1,3 +1,0 @@
-displayName = Depth of Time
-author = Ren
-version = 0.1
