+++ Items/Placeable/Backpack.png
